Moves dtos to respective providers
