+++ opencdms/dtos/station_schema.py
