+++ opencdms/provider/clide/dtos/station.py
