+++ opencdms/provider/clide/dtos/stationstatu.py
