+++ opencdms/provider/climsoft/__init__.py
@@ -31,7 +31,7 @@ import logging
 from types import ModuleType
 from opencdms.provider.base import CDMSProvider
 from opencdms.provider.climsoft.models import v4_1_1_core as climsoft_models
-from opencdms.dtos import climsoft as climsoft_schemas
+from opencdms.provider.climsoft import dtos as climsoft_schemas
 
 LOGGER = logging.getLogger(__name__)
