+++ opencdms/provider/midas_pg/__init__.py
@@ -34,7 +34,7 @@ import os
 import logging
 from types import ModuleType
 from opencdms.provider.midas_pg.models import core as midas_models
-from opencdms.dtos import midas as midas_schemas
+from opencdms.provider.midas_pg import dtos as midas_schemas
 
 from opencdms.provider.base import CDMSProvider
 from opencdms.fileformats.text import read_badc
