+++ tests/integration/providers/test_mch_english.py
@@ -2,7 +2,7 @@ import pytest
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
-from opencdms.dtos.mch import station as mch_station_schema
+from opencdms.provider.mch.dtos import station as mch_station_schema
 from opencdms.provider.mch.models import english as mch_english
 from opencdms.provider.mch import MCHProvider
 from opencdms.utils.db import get_mch_english_connection_string
