+++ tests/integration/providers/test_midas_pg.py
@@ -4,7 +4,7 @@ from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy.sql import text as sa_text
 
-from opencdms.dtos.midas import source as midas_source_schema
+from opencdms.provider.midas_pg.dtos import source as midas_source_schema
 from opencdms.provider.midas_pg.models import core as midas_models
 from opencdms.provider.midas_pg import MidasPgProvider
 from opencdms.utils.db import get_midas_pg_connection_string
