+++ tests/integration/providers/test_opencdms.py
@@ -1,10 +1,10 @@
 from sqlalchemy import create_engine
 from sqlalchemy.sql import text as sa_text
 
-from opencdms.dtos.clide import station as clide_station
-from opencdms.dtos.clide import stationstatu as clide_station_status
-from opencdms.dtos.clide import stationtimezone as clide_station_timezone
-from opencdms.dtos.mch import station as mch_station
+from opencdms.provider.clide.dtos import station as clide_station
+from opencdms.provider.clide.dtos import stationstatu as clide_station_status
+from opencdms.provider.clide.dtos import stationtimezone as clide_station_timezone
+from opencdms.provider.mch.dtos import station as mch_station
 import opencdms.provider.clide.models as clide
 from opencdms.provider.mch.models import english as mch
 from opencdms.provider.opencdms import OpenCDMSProvider, ProviderConfig
